+++ lib/Cauldron/Managed.hs
@@ -1,10 +1,15 @@
 {-# LANGUAGE BlockArguments #-}
 
+-- | A datatype that encapsulates @with...@-style callback-taking functions that
+-- acquire and release resources.
+--
+-- Here's [a video about how it works](https://youtu.be/2v7BAQh_dRs). 
 module Cauldron.Managed
   ( -- * The Managed monad for handling resources
     Managed,
     managed,
     with,
+    runManaged,
   )
 where
 
@@ -50,10 +55,6 @@ instance MonadFix Managed where
     where
       unManage (Managed a) = a
 
--- | Make use of the managed resource by supplying a callback.
-with :: Managed a -> (a -> IO b) -> IO b
-with (Managed r) = r
-
 instance Functor Managed where
   fmap f (Managed m) = Managed (\k -> m (\x -> k (f x)))
   {-# INLINE fmap #-}
@@ -80,3 +81,17 @@ instance MonadFail Managed where
     fail s = Managed (\return_ -> do
         a <- fail @IO s
         return_ a )
+
+instance Semigroup a => Semigroup (Managed a) where
+    (<>) = liftA2 (<>)
+
+instance Monoid a => Monoid (Managed a) where
+    mempty = pure mempty
+
+-- | Make use of the managed resource by supplying a callback.
+with :: Managed a -> (a -> IO b) -> IO b
+with (Managed r) = r
+
+-- | Run a `Managed` computation, enforcing that no acquired resources leak
+runManaged :: Managed () -> IO ()
+runManaged (Managed r) = r pure